+++ frontend/src/api/ApiAuthService.ts
@@ -27,7 +27,6 @@ class ApiAuthService extends ApiService {
   }): Promise<AuthResponse> {
     return await this.request<AuthResponse>("POST", `${this.CRUD_PATH}/`, {
       data: auth,
-      asFormData: true,
     });
   }